Function Cleanup - Bollinger Bands

+++ TradingResearchAPI/TradingResearchAPI/Models/TradeRecommendation.cs
@@ -10,9 +10,9 @@ namespace TradingResearchAPI.Models
         public string Symbol { get; set; }
         public TradeAction TradeAction { get; set; }
         public String TradeActionDescription { get; set; }
-        public double PurchaseRecommendation { get; set; }
-        public double SellRecommendation { get; set; }
-        public double percentReturnPotential { get; set; }
+        public decimal PurchaseRecommendation { get; set; }
+        public decimal SellRecommendation { get; set; }
+        public decimal percentReturnPotential { get; set; }
         public Candle MostRecentTradingSession { get; set; }
